requirements.txt
cloudinary==1.15.0
Django>=2.0
django-cloudinary-storage==0.2.3
Pillow==5.3.0
djangorestframework